rename : aws s3 package 위치 변경 (domain > global)
