--- src/main/java/com/crude/travelcrew/domain/awss3/constants/FileExtension.java
@@ -1,5 +1,0 @@
-package com.crude.travelcrew.domain.awss3.constants;
-
-public enum FileExtension {
-	JPG, JPEG, PNG
-}
