+++ src/main/java/com/crude/travelcrew/global/awss3/constants/FileExtension.java
@@ -1,0 +1,5 @@
+package com.crude.travelcrew.global.awss3.constants;
+
+public enum FileExtension {
+	JPG, JPEG, PNG
+}
